Share project "ScrollbarTest1" into "https://se-project-group8.googlecode.com/svn"
Please ignore the title - I was originally naming it for scrollbar, but we don't need that anymore. It is a zoom file that is supposed to work with an actual device, but I can't test that. I know that it compiles and runs without errors, and my image appeared on the screen and I could drag it around, but that's all I know that it does.