Major revision. See below to read. ----------------------------------------------------- Sms app can now be set as the default app. When it is the default app. It has permissions to read, write, and delete from the stock sms content://sms database and also is in charge of managing that database on send and recieve.
In order to set it as the default app, press the button at the bottom of the main landing page.

Drafts can now be saved when set as the stock default app. Currently there is no safeguard from preventing someone from clicking on save draft when the app is not the default app. There needs to be one implemented using Utils.isDefaultSmsApp(context), which returns true iff we are the default app. The warning should make it to where the button does nothing expect have Toast display a message such as "Please set this app as the default SMS app first." You can save as many drafts as you want. Saving multiple drafts for a single phone number is allowed. 

To be implemented (Regarding drafts):
Draft inbox (Similar to Conversation, just with a filter to only add if there is a draft, will also include searching).
EditDraft by clicking on draft name inside inbox, conversation, or draft inbox.
Automatic deletion of draft on click, ie whenever the draft is brought into editing.
Overriding onPause() (called whenever a new activity is brought to front and before onStop/onDestroy()) to save message and number currently inputed as a draft.

Other changes/noteworthy side effects:
receivers for inbox and conversation now are closed properly whenever the app closes
reverted back to old algorithm for replacing contact name
autoreply does not populate in the inbox whenever we are the default sms app. this is intentional (think autoreply should be set to only be on if we are the default sms app.
Min api raised to 19. This is due to major changes in android sms management since api 4.4. See here for more info: https://www.youtube.com/watch?v=mdq0R2WQssQ